move theme, router and extensions under `utils` folder
