+++ lib/runner_io.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:intl/intl_standalone.dart';
 
-import 'src/core/service_locator/service_locator.dart';
+import 'src/core/utils/service_locator.dart';
 import 'src/core/widget/app.dart';
 
 /// App running on IO device (mobile)
