+++ lib/runner_web.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:intl/intl_browser.dart';
 
-import 'src/core/service_locator/service_locator.dart';
+import 'src/core/utils/service_locator.dart';
 import 'src/core/widget/app.dart';
 
 /// App running on web (html)
