+++ lib/src/core/utils/extensions/string_extensions.dart
