+++ lib/src/core/utils/router/router.dart
@@ -1,8 +1,8 @@
 import 'package:auto_route/annotations.dart';
 
-import '../../features/cart/presentation/widget/cart_screen.dart';
-import '../../features/item_details/presentation/widget/item_details_screen.dart';
-import '../../features/navigation/presentation/widget/navigation_screen.dart';
+import '../../../features/cart/presentation/widget/cart_screen.dart';
+import '../../../features/item_details/presentation/widget/item_details_screen.dart';
+import '../../../features/navigation/presentation/widget/navigation_screen.dart';
 
 
 @MaterialAutoRouter(
