+++ lib/src/features/cart/domain/usecases/get_cart.dart
@@ -1,4 +1,4 @@
-import '../../../../core/service_locator/service_locator.dart';
+import '../../../../core/utils/service_locator.dart';
 import '../../data/repository/cart_repository.dart';
 import '../models/cart_content/cart_content.dart';
