+++ lib/src/features/item_details/data/repository/item_details_repository.dart
@@ -1,4 +1,4 @@
-import '../../../../core/service_locator/service_locator.dart';
+import '../../../../core/utils/service_locator.dart';
 import '../../domain/models/details_model.dart';
 import '../../domain/repository/item_details_repository.dart';
 import '../datasource/dto_get_item_details.dart';
