+++ lib/src/features/item_details/domain/usecases/get_item_details_from_server.dart
@@ -1,4 +1,4 @@
-import '../../../../core/service_locator/service_locator.dart';
+import '../../../../core/utils/service_locator.dart';
 import '../../data/repository/item_details_repository.dart';
 import '../models/details_model.dart';
