+++ lib/src/features/store/domain/usecases/get_store_items.dart
@@ -1,4 +1,4 @@
-import '../../../../core/service_locator/service_locator.dart';
+import '../../../../core/utils/service_locator.dart';
 import '../../data/repository/store_repository.dart';
 import '../models/store_items/store_items.dart';
